@@ -1,10 +1,11 @@
-import {Injectable} from "@angular/core";
-import {Http, Headers, Response} from "@angular/http";
-import {User} from "./user";
-import {Config} from "../config";
-import {Observable} from "rxjs/Rx";
+import { Injectable } from "@angular/core";
+import { Http, Headers, Response } from "@angular/http";
+import { Observable } from "rxjs";
 import "rxjs/add/operator/do";
 import "rxjs/add/operator/map";
+
+import { User } from "./user";
+import { Config } from "../config";
 
 @Injectable()
 export class UserService {
@@ -60,7 +61,7 @@ export class UserService {
     .catch(this.handleErrors);
   }
 
-  handleErrors(error: Response) {
+  private handleErrors(error: Response) {
     console.log(JSON.stringify(error.json()));
     return Observable.throw(error);
   }
