@@ -1,9 +1,11 @@
-export class GroceriesPage {
+import { browser, element, by } from 'protractor';
+
+export class SampleGroceriesPage {
   navigateTo() {
     return browser.get('/');
   }
 
   getParagraphText() {
-    return element(by.css('groceries-app h1')).getText();
+    return element(by.css('app-root h1')).getText();
   }
 }
