@@ -1,9 +1,12 @@
-import {bootstrap} from "@angular/platform-browser-dynamic";
-import {enableProdMode} from "@angular/core";
-import {GroceriesAppComponent, environment} from "./app/";
+import './polyfills.ts';
+
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { enableProdMode } from '@angular/core';
+import { environment } from './environments/environment';
+import { AppModule } from './app/';
 
 if (environment.production) {
   enableProdMode();
 }
 
-bootstrap(GroceriesAppComponent);
+platformBrowserDynamic().bootstrapModule(AppModule);
