@@ -1,2 +1,2 @@
-/// <reference path="../typings/browser.d.ts" />
-declare var module: { id: string };
+// Typings reference file, you can add your own global typings here
+// https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
